# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Esteticar Car Wash Booking Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface MessageResponse {
  message: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  customer: "customer",
  provider: "provider",
  admin: "admin",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  phone?: string | null;
  locationId?: number | null;
  createdAt: string;
}

export interface RegisterBody {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  phone?: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export interface Service {
  id: number;
  name: string;
  description: string;
  price: number;
  durationMinutes: number;
  imageUrl?: string | null;
  isActive: boolean;
  createdAt: string;
}

export interface CreateServiceBody {
  name: string;
  description: string;
  price: number;
  durationMinutes: number;
  imageUrl?: string;
  isActive?: boolean;
}

export interface Location {
  id: number;
  name: string;
  address: string;
  city: string;
  lat?: number | null;
  lng?: number | null;
  phone?: string | null;
  openTime: string;
  closeTime: string;
  isActive: boolean;
  createdAt: string;
}

export interface CreateLocationBody {
  name: string;
  address: string;
  city: string;
  lat?: number;
  lng?: number;
  phone?: string;
  openTime: string;
  closeTime: string;
  isActive?: boolean;
}

export interface Timeslot {
  id: number;
  locationId: number;
  date: string;
  startTime: string;
  endTime: string;
  isAvailable: boolean;
  maxBookings: number;
  currentBookings: number;
}

export interface ConfigureTimeslotsBody {
  locationId: number;
  startDate: string;
  endDate: string;
  openTime: string;
  closeTime: string;
  slotDurationMinutes: number;
  maxBookingsPerSlot: number;
}

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  confirmed: "confirmed",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Booking {
  id: number;
  userId: number;
  serviceId: number;
  locationId: number;
  timeslotId: number;
  status: BookingStatus;
  date: string;
  startTime: string;
  endTime: string;
  notes?: string | null;
  service?: Service | null;
  location?: Location | null;
  user?: User | null;
  createdAt: string;
}

export interface CreateBookingBody {
  serviceId: number;
  locationId: number;
  timeslotId: number;
  date: string;
  notes?: string;
}

export type UpdateBookingStatusBodyStatus =
  (typeof UpdateBookingStatusBodyStatus)[keyof typeof UpdateBookingStatusBodyStatus];

export const UpdateBookingStatusBodyStatus = {
  confirmed: "confirmed",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateBookingStatusBody {
  status: UpdateBookingStatusBodyStatus;
}

export interface CreateProviderBody {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  phone?: string;
  locationId?: number;
}

export interface AssignProviderBody {
  locationId: number;
}

export interface ProviderStats {
  totalCompleted: number;
  totalToday: number;
  totalThisWeek: number;
  totalThisMonth: number;
}

export interface AdminDashboard {
  totalBookings: number;
  todayBookings: number;
  completedBookings: number;
  cancelledBookings: number;
  confirmedBookings: number;
  totalRevenue: number;
  totalLocations: number;
  totalProviders: number;
  totalCustomers: number;
  recentBookings: Booking[];
}

export interface CustomerDashboard {
  upcomingCount: number;
  completedCount: number;
  nextBooking?: Booking | null;
  recentBookings: Booking[];
}

export type ListTimeslotsParams = {
  locationId: number;
  date: string;
  serviceId?: number;
};

export type ListBookingsParams = {
  status?: ListBookingsStatus;
  upcoming?: boolean;
};

export type ListBookingsStatus =
  (typeof ListBookingsStatus)[keyof typeof ListBookingsStatus];

export const ListBookingsStatus = {
  confirmed: "confirmed",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type GetProviderScheduleParams = {
  date?: string;
};

export type ListAllBookingsParams = {
  status?: ListAllBookingsStatus;
  locationId?: number;
  date?: string;
};

export type ListAllBookingsStatus =
  (typeof ListAllBookingsStatus)[keyof typeof ListAllBookingsStatus];

export const ListAllBookingsStatus = {
  confirmed: "confirmed",
  completed: "completed",
  cancelled: "cancelled",
} as const;