# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Esteticar Car Wash Booking Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new customer account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current logged-in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # SERVICES
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings for current user (or all if admin)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [confirmed, completed, cancelled]
        - name: upcoming
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a new booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Slot unavailable or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a specific booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateBookingStatus
      tags: [bookings]
      summary: Update booking status (cancel, complete)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingStatusBody"
      responses:
        "200":
          description: Booking updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  # PROVIDER ROUTES
...
  /admin/providers:
    get:
      operationId: listProviders
      tags: [admin]
      summary: List all service providers
      responses:
        "200":
          description: List of providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createProvider
      tags: [admin]
      summary: Create a service provider account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProviderBody"
      responses:
        "201":
          description: Provider created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/providers/{id}/assign:
    post:
      operationId: assignProviderToLocation
      tags: [admin]
      summary: Assign provider to a location
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignProviderBody"
      responses:
        "200":
          description: Provider assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/bookings:
    get:
      operationId: listAllBookings
      tags: [admin]
      summary: List all bookings (admin view)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [confirmed, completed, cancelled]
        - name: locationId
          in: query
          required: false
          schema:
            type: integer
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        "200":
          description: All bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [customer, provider, admin]
        phone:
          type: string
          nullable: true
        locationId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - createdAt
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        phone:
          type: string
      required:
        - name
        - email
        - password
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required:
        - user
        - message
    Service:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        durationMinutes:
          type: integer
        imageUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - price
        - durationMinutes
        - isActive
        - createdAt
...
    Location:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
        city:
          type: string
        lat:
          type: number
          nullable: true
        lng:
          type: number
          nullable: true
        phone:
          type: string
          nullable: true
        openTime:
          type: string
        closeTime:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - address
        - city
        - openTime
        - closeTime
        - isActive
        - createdAt
...
    Booking:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        serviceId:
          type: integer
        locationId:
          type: integer
        timeslotId:
          type: integer
        status:
          type: string
          enum: [confirmed, completed, cancelled]
        date:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
        notes:
          type: string
          nullable: true
        service:
          $ref: "#/components/schemas/Service"
          nullable: true
        location:
          $ref: "#/components/schemas/Location"
          nullable: true
        user:
          $ref: "#/components/schemas/User"
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - serviceId
        - locationId
        - timeslotId
        - status
        - date
        - startTime
        - endTime
        - createdAt
    CreateBookingBody:
      type: object
      properties:
        serviceId:
          type: integer
        locationId:
          type: integer
        timeslotId:
          type: integer
        date:
          type: string
          format: date
        notes:
          type: string
      required:
        - serviceId
        - locationId
        - timeslotId
        - date
    UpdateBookingStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [confirmed, completed, cancelled]
      required:
        - status
    CreateProviderBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        phone:
          type: string
        locationId:
          type: integer
      required:
        - name
        - email
        - password
    AssignProviderBody:
      type: object
      properties:
        locationId:
          type: integer
      required:
        - locationId
...