# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Esteticar Car Wash Booking Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from "./bookingStatus";
import type { Location } from "./location";
import type { Service } from "./service";
import type { User } from "./user";

export interface Booking {
  id: number;
  userId: number;
  serviceId: number;
  locationId: number;
  timeslotId: number;
  status: BookingStatus;
  date: Date;
  startTime: string;
  endTime: string;
  notes?: string | null;
  service?: Service | null;
  location?: Location | null;
  user?: User | null;
  createdAt: Date;
}